Makefile and General Library Improvements
- Cleaned up Makefile and updated -I flags to use -isystem to ignore warnings for external libraries. 
- Updated metajson to the latest version (Library moved to lithium_json). Updated library usage. 
- Added link dependency on boost_system. 
- Updated google_test include pathing.

+++ src/Interfaces/BootstrapCache.hpp
@@ -35,8 +35,8 @@ public:
         Endpoints::TechnologyType technology,
         OneEndpointBootstrapReadFunction const& callback) const = 0;
 
-    virtual std::uint32_t CachedBootstrapCount() const = 0;
-    virtual std::uint32_t CachedBootstrapCount(Endpoints::TechnologyType technology) const = 0;
+    virtual std::size_t CachedBootstrapCount() const = 0;
+    virtual std::size_t CachedBootstrapCount(Endpoints::TechnologyType technology) const = 0;
 };
 
 //------------------------------------------------------------------------------------------------
